Clean up 'hashTable' files - now only one (correct) set of hT files
TL;DR - deleted files with incorrect solutions. The 'hashTable'
files in this commit are correct; disregard the various multiples
of 'hashTable'-related files in our previous commits.

------------

Our original attempt at solving 'hashTable' were correct, but
we edited the hash table helper functions while not realizing
that we weren't supposed to. These initial solutions were in
files called 'hashTable.js' and 'hashTableHelpers.js'

We refactored our solution completely to accomodate the rule
that no changes should be made to the hash table helpers, except
in the instance of bugs. This second set of solutions were created
in the 'baseHashTable.js' and 'baseHashTableHelpers.js' files. We
didn't want to delete the original files until we were sure that
the second solution passed all of the tests, including the two
advanced ones regarding doubling/halving the table.

After we completed our second set of solutions, we committed
all files - including the original set ('hashTable.js' and
 'hashTableHelpers.js'), and the second set ('baseHashTable.js'
 and 'baseHashTableHelpers.js').

This commit is the final commit to date - we deleted the files
containing our original (incorrect) solution, leaving only the
second set of files. Finally, we renamed those two files to have
the original names of 'hashTable.js' and 'hashTableHelpers.js'.
The files in this commit have our single (correct) solution.